requirements.txt
asgiref==3.8.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
psycopg2-binary==2.9.8
debugpy
drf-nested-routers
